Add handling of duplicate patients to backend and frontend.
A duplicate is identified by firstname and surname.
When a patient is created it is checked if this combination of first- and surname is already existing. A excaption is raised an caught by the appropriate ExceptionTranslator, which creates the HTTP response. This response is caught in the frontend and a message is displayed.