Change handling of patient duplicate error. Disable default error handling for this error.
The patient duplicate error is send as a http 400 response.
By default every these will trigger the alert-error.component and a message will be displayed for 5 seconds. This is undesired for the patient duplicate error, as a unique message is displayed. Instead of removing the default error handling from the patient creation, the default error handling has been altered and will not be triggered by a patient duplicate error.